ev-view: avoid view events while view is still loading
This is a small optimization to avoid processing
mouseover events over a blank view while the view
is still loading.

While working on a pdf that takes some time to load
I noticed that moving the mouse over the blank view
created calls to eg. check if an annotation was
under the mouse pointer, all that is wasted cycles
until the file is finally loaded.